Remove core server file.
