+++ rasa/core/restore.py
@@ -142,61 +142,3 @@ def load_tracker_from_json(tracker_dump: Text, domain: Domain) -> DialogueStateT
     return DialogueStateTracker.from_dict(
         sender_id, tracker_json.get("events", []), domain.slots
     )
-
-
-async def serve_application(
-    core_model: Text = None,
-    nlu_model: Optional[Text] = None,
-    tracker_dump: Optional[Text] = None,
-    port=constants.DEFAULT_SERVER_PORT,
-    enable_api=True,
-    endpoints=None,
-):
-    input_channels = run.create_http_input_channels("cmdline", None)
-
-    app = run.configure_app(input_channels, enable_api=enable_api, port=port)
-
-    logger.info(
-        "Starting Rasa Core server on "
-        "{}".format(constants.DEFAULT_SERVER_FORMAT.format(port))
-    )
-
-    # noinspection PyShadowingNames
-    async def load_agent_and_tracker(app, loop):
-        agent = await run.load_agent_on_start(
-            core_model, endpoints, nlu_model, app, loop
-        )
-
-        tracker = load_tracker_from_json(tracker_dump, agent.domain)
-        await replay_events(tracker, agent)
-
-    app.register_listener(load_agent_and_tracker, "before_server_start")
-    app.run(host="0.0.0.0", port=port, access_log=logger.isEnabledFor(logging.DEBUG))
-
-
-if __name__ == "__main__":
-    # Running as standalone python application
-    arg_parser = create_argument_parser()
-    cmdline_args = arg_parser.parse_args()
-
-    rasa.utils.io.configure_colored_logging(cmdline_args.loglevel)
-    _endpoints = AvailableEndpoints.read_endpoints(cmdline_args.endpoints)
-
-    print (
-        cliutils.wrap_with_color(
-            "We'll recreate the dialogue state. After that you can chat "
-            "with the bot, continuing the input conversation.",
-            rasa.cli.utils.bcolors.OKGREEN + rasa.cli.utils.bcolors.UNDERLINE,
-        )
-    )
-
-    _loop = asyncio.get_event_loop()
-    _loop.run_until_complete(
-        serve_application(
-            cmdline_args.core,
-            cmdline_args.nlu,
-            cmdline_args.port,
-            cmdline_args.enable_api,
-            _endpoints,
-        )
-    )
