+++ tests/core/test_evaluation.py
@@ -1,6 +1,5 @@
 import os
 
-from rasa.core.server import nlu_model_and_evaluation_files_from_archive
 from rasa.core.test import _generate_trackers, collect_story_predictions, test
 from rasa.model import add_evaluation_file_to_model
 
@@ -69,24 +68,3 @@ async def test_end_to_end_evaluation_script_unknown_entity(tmpdir, default_agent
     assert story_evaluation.evaluation_store.has_prediction_target_mismatch()
     assert len(story_evaluation.failed_stories) == 1
     assert num_stories == 1
-
-
-async def test_stack_model_intent_evaluation(
-    tmpdir, trained_stack_model, default_nlu_data
-):
-    with open(default_nlu_data, "r") as f:
-        nlu_data = f.read()
-
-    # add evaluation data to model archive
-    new_model_path = add_evaluation_file_to_model(
-        trained_stack_model, nlu_data, data_format="md"
-    )
-
-    nlu_model_path, nlu_files = await nlu_model_and_evaluation_files_from_archive(
-        new_model_path, tmpdir
-    )
-
-    assert len(nlu_files) == 1
-    evaluation = run_evaluation(nlu_files[0], nlu_model_path)
-
-    assert set(evaluation.keys()) == {"intent_evaluation", "entity_evaluation"}
